# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pywel Knowledge Standard API",
    "version": "1.0.0",
    "description": "Unofficial read-only local Crimson Desert reference interface. Frozen v1 agent contract. Historical coverage is finite; the latest indexed patch is not the current game patch. No hosted endpoint is promised. All inputs are closed: unknown or duplicate parameters, unsupported enum values, and out-of-range numbers are rejected.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local loopback reference server; PORT can select another local port."
    }
  ],
...
  "paths": {
...
    "/v1": {
      "get": {
        "operationId": "getServiceDescriptor",
        "tags": [
          "discovery"
        ],
        "summary": "Read the local service descriptor and supported endpoint map",
        "responses": {
          "200": {
            "description": "Versioned getServiceDescriptor response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/agent-contract.schema.json#/definitions/getServiceDescriptor"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": []
      }
    },
    "/v1/codebook": {
      "get": {
        "operationId": "getCompactCodebook",
        "tags": [
          "discovery"
        ],
        "summary": "Decode compact evidence packet arrays",
        "responses": {
          "200": {
            "description": "Versioned getCompactCodebook response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/agent-contract.schema.json#/definitions/getCompactCodebook"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": []
      }
    },
    "/v1/search": {
      "get": {
        "operationId": "searchKnowledge",
        "tags": [
          "knowledge"
        ],
        "summary": "Search entities, claims, and strategies",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 500
            },
            "required": true
          },
          {
            "$ref": "#/components/parameters/Patch"
          },
          {
            "$ref": "#/components/parameters/Platform"
          },
          {
            "$ref": "#/components/parameters/Locale"
          },
          {
            "$ref": "#/components/parameters/Spoiler"
          },
          {
            "$ref": "#/components/parameters/IncludeRetracted"
          },
          {
            "$ref": "#/components/parameters/IncludeSuperseded"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit100"
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned searchKnowledge response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/agent-contract.schema.json#/definitions/searchKnowledge"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "description": "Deterministic ranked search. All list pagination uses offset and next_offset, scoped to one build and unchanged filters."
      }
    },
    "/v1/entities": {
      "get": {
        "operationId": "listEntities",
        "tags": [
          "knowledge"
        ],
        "summary": "List or search canonical entities",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 500
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EntityType"
            }
          },
          {
            "name": "subtype",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 80,
              "enum": [
                "title",
                "combat",
                "progression",
                "inventory",
                "crafting",
                "navigation",
                "interface",
                "performance",
                "economy",
                "movement",
                "system",
                "mechanic",
                "weapon",
                "armor",
                "accessory",
                "tool",
                "mount_tack",
                "consumable",
                "material",
                "quest_item",
                "abyss_gear",
                "currency",
                "item",
                "main",
                "faction",
                "side",
                "bounty",
                "tutorial",
                "challenge",
                "quest",
                "continent",
                "region",
                "settlement",
                "stronghold",
                "camp",
                "landform",
                "ruin",
                "facility",
                "landmark",
                "spire",
                "interior",
                "service",
                "shop",
                "dungeon",
                "realm",
                "location",
                "player_character",
                "npc",
                "boss",
                "mount",
                "companion",
                "enemy",
                "character",
                "active",
                "passive",
                "skill",
                "buff",
                "debuff",
                "immunity",
                "effect",
                "utility",
                "major",
                "minor",
                "institution",
                "cooking",
                "upgrade",
                "recipe",
                "resource",
                "build",
                "life_skill",
                "dispatch"
              ]
            },
            "description": "A subtype from the shipped entity-subtypes.json vocabulary; unknown subtypes are rejected."
          },
          {
            "name": "include_redirects",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit100"
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned listEntities response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/agent-contract.schema.json#/definitions/listEntities"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/entities/{id}": {
      "get": {
        "operationId": "getEntityBundle",
        "tags": [
          "knowledge"
        ],
        "summary": "Get an entity plus context-matching claims, strategies, and evidence",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 200,
              "pattern": "^[A-Za-z0-9._-]+$"
            },
            "required": true,
            "description": "Stable entity ID or slug."
          },
          {
            "$ref": "#/components/parameters/Patch"
          },
          {
            "$ref": "#/components/parameters/Platform"
          },
          {
            "$ref": "#/components/parameters/Locale"
          },
          {
            "$ref": "#/components/parameters/Spoiler"
          },
          {
            "$ref": "#/components/parameters/IncludeRetracted"
          },
          {
            "$ref": "#/components/parameters/IncludeSuperseded"
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned getEntityBundle response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/agent-contract.schema.json#/definitions/getEntityBundle"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/entities/{id}/relationships": {
      "get": {
        "operationId": "getEntityRelationships",
        "tags": [
          "knowledge"
        ],
        "summary": "Traverse entity-valued atomic claims",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 200,
              "pattern": "^[A-Za-z0-9._-]+$"
            },
            "required": true,
            "description": "Stable entity ID or slug."
          },
          {
            "$ref": "#/components/parameters/Patch"
          },
          {
            "$ref": "#/components/parameters/Platform"
          },
          {
            "$ref": "#/components/parameters/Locale"
          },
          {
            "$ref": "#/components/parameters/Spoiler"
          },
          {
            "$ref": "#/components/parameters/IncludeRetracted"
          },
          {
            "$ref": "#/components/parameters/IncludeSuperseded"
          },
          {
            "name": "depth",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 3,
              "default": 1
            }
          },
          {
            "$ref": "#/components/parameters/Limit500"
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned getEntityRelationships response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/agent-contract.schema.json#/definitions/getEntityRelationships"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/claims": {
      "get": {
        "operationId": "listClaims",
        "tags": [
          "knowledge"
        ],
        "summary": "Retrieve atomic claims under an explicit version context",
        "parameters": [
          {
            "name": "subject",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^ent_[a-z0-9]{12,64}$"
            }
          },
          {
            "name": "predicate",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ClaimStatus"
            }
          },
          {
            "$ref": "#/components/parameters/Patch"
          },
          {
            "$ref": "#/components/parameters/Platform"
          },
          {
            "$ref": "#/components/parameters/Locale"
          },
          {
            "$ref": "#/components/parameters/Spoiler"
          },
          {
            "$ref": "#/components/parameters/IncludeRetracted"
          },
          {
            "$ref": "#/components/parameters/IncludeSuperseded"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit200"
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned listClaims response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/agent-contract.schema.json#/definitions/listClaims"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/answer": {
      "get": {
        "operationId": "answerQuestion",
        "tags": [
          "knowledge"
        ],
        "summary": "Build an evidence packet for a Crimson Desert question",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 1000
            },
            "required": true
          },
          {
            "$ref": "#/components/parameters/Patch"
          },
          {
            "$ref": "#/components/parameters/Platform"
          },
          {
            "$ref": "#/components/parameters/Locale"
          },
          {
            "$ref": "#/components/parameters/Spoiler"
          },
          {
            "$ref": "#/components/parameters/IncludeRetracted"
          },
          {
            "$ref": "#/components/parameters/IncludeSuperseded"
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "compact",
                "full"
              ],
              "default": "compact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Versioned answerQuestion response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/agent-contract.schema.json#/definitions/answerQuestion"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "description": "Default compact v3 contains complete typed assertions and evidence indexes. Full v1 retains at most 20 claims and 8 strategies; compact retains at most 5 claims and 2 strategies. selection counts disclose matched and returned records. Omitted stale/disputed evidence still limits answer state. The full_path is a fuller bounded projection, not a guarantee of completeness."
      }
    },
...
    "/v1/patches": {
      "get": {
        "operationId": "listPatches",
        "tags": [
          "sources"
        ],
        "summary": "List known patches and their official evidence",
        "responses": {
          "200": {
            "description": "Versioned listPatches response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/agent-contract.schema.json#/definitions/listPatches"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "414": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": []
      }
    },
...
  },
  "components": {
    "parameters": {
      "Patch": {
        "name": "patch",
        "in": "query",
        "schema": {
          "type": "string",
          "pattern": "^[0-9]+(?:\\.[0-9]+){1,3}$",
          "maxLength": 32
        },
        "description": "Omit to use and report the latest indexed historical patch. Syntactically valid unindexed versions remain explicit unknown applicability."
      },
      "Platform": {
        "name": "platform",
        "in": "query",
        "schema": {
          "$ref": "#/components/schemas/Platform",
          "default": "all"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid, duplicate, unknown, unsupported, or out-of-range input.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/agent-contract.schema.json#/definitions/error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Route or record not found, or strategy outside the requested context.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/agent-contract.schema.json#/definitions/error"
            }
          }
        }
      },
      "RequestTooLarge": {
        "description": "Request URL exceeds 8192 characters.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/agent-contract.schema.json#/definitions/error"
            }
          }
        }
      },
      "InternalError": {
        "description": "Request could not be completed; internal details are withheld.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/agent-contract.schema.json#/definitions/error"
            }
          }
        }
      }
    },
...
  },
...
}